refactor(architecture-diagram): extract model and spec modules
Move diagnostics, authored models, YAML parsing, and spec comparison behind owner modules while preserving renderer behavior.

Stack-Id: architecture-diagram-m1-3d8ce7
Stack-Position: 1/4
