+++ skills/architecture-diagram/engine/render.py
@@ -32,10 +32,23 @@ from math import hypot
 from pathlib import Path
 from typing import Protocol
 
-import yaml
 
 from . import fetch_icons
 from .data import data_path
+
+from .compare import compare_specs
+from .diagnostics import (
+    QUALITY_PROFILES,
+    QUALITY_STANDARD,
+    SEVERITY_ERROR,
+    SEVERITY_WARNING,
+    Diagnostic,
+    apply_quality_profile,
+    count_by_severity,
+    suppress_derived,
+)
+from .model import Edge, Node, Spec, Zone
+from .spec import SpecError, _spec_error, load_spec
 
 # --- layout constants -------------------------------------------------------
 
@@ -120,533 +133,14 @@ def _escape(text: str) -> str:
 
 # --- diagnostics -------------------------------------------------------------
 
-SEVERITY_ERROR = "error"
-SEVERITY_WARNING = "warning"
-_SEVERITIES = (SEVERITY_ERROR, SEVERITY_WARNING)
-
-QUALITY_STANDARD = "standard"
-QUALITY_SHOWCASE = "showcase"
-QUALITY_PROFILES = (QUALITY_STANDARD, QUALITY_SHOWCASE)
 
 # Codes in this namespace describe how the drawing reads rather than whether
 # the spec is answerable: two routes crossing, a label sitting on a route.
 # They are warnings under the standard profile and errors under showcase, so
 # stricter geometry rules can ship without breaking specs that already render.
-PROFILE_SENSITIVE_NAMESPACE = "composition/"
-
-DIAGNOSTIC_SCHEMA_VERSION = 1
-
-
-@dataclass
-class Diagnostic:
-    """One machine-readable finding.
-
-    The predecessor of this type was `list[str]` of English prose printed to
-    stderr, which forced the calling agent to regex sentences and left it to
-    invent its own repair — usually by editing the SVG, which is exactly the
-    move that destroys the reproducibility this renderer exists to provide.
-    `code` is the stable identity, `subject` says what the finding is about,
-    `evidence` carries the numbers needed to locate it, and
-    `supported_fixes` bounds the repair to changes an author can make in the
-    spec.
-    """
-
-    code: str
-    severity: str
-    message: str
-    subject: dict[str, object] = field(default_factory=dict)
-    evidence: dict[str, object] = field(default_factory=dict)
-    supported_fixes: tuple[str, ...] = ()
-    # Codes this finding makes redundant. A derived diagnostic reported
-    # alongside its cause sends the agent chasing symptoms.
-    suppresses: tuple[str, ...] = ()
-
-    def __post_init__(self) -> None:
-        if self.severity not in _SEVERITIES:
-            raise ValueError(
-                f"diagnostic {self.code!r} has unknown severity {self.severity!r}; "
-                f"expected one of {_SEVERITIES}"
-            )
-
-    def to_dict(self) -> dict[str, object]:
-        payload: dict[str, object] = {
-            "code": self.code,
-            "severity": self.severity,
-            "message": self.message,
-            "subject": dict(self.subject),
-            "evidence": dict(self.evidence),
-            "supported_fixes": list(self.supported_fixes),
-        }
-        if self.suppresses:
-            payload["suppresses"] = list(self.suppresses)
-        return payload
-
-
-def apply_quality_profile(
-    diagnostics: list[Diagnostic], quality: str
-) -> list[Diagnostic]:
-    """Raise profile-sensitive findings to errors under the showcase profile."""
-    if quality not in QUALITY_PROFILES:
-        raise ValueError(
-            f"unknown quality profile {quality!r}; expected one of {QUALITY_PROFILES}"
-        )
-    if quality != QUALITY_SHOWCASE:
-        return list(diagnostics)
-    out = []
-    for d in diagnostics:
-        if (
-            d.code.startswith(PROFILE_SENSITIVE_NAMESPACE)
-            and d.severity != SEVERITY_ERROR
-        ):
-            out.append(
-                Diagnostic(
-                    code=d.code,
-                    severity=SEVERITY_ERROR,
-                    message=d.message,
-                    subject=d.subject,
-                    evidence=d.evidence,
-                    supported_fixes=d.supported_fixes,
-                    suppresses=d.suppresses,
-                )
-            )
-        else:
-            out.append(d)
-    return out
-
-
-def suppress_derived(diagnostics: list[Diagnostic]) -> list[Diagnostic]:
-    """Drop findings a reported cause makes redundant.
-
-    Suppression is one level deep and keyed on `code`: a record is dropped
-    when any record in the input names its code, and a dropped record still
-    suppresses. Resolving chains instead has no well-founded answer — if A
-    suppresses B and C suppresses A, then removing B leaves it hidden with
-    no visible cause, while keeping B oscillates on the next pass, and a
-    mutual pair has no defensible winner at all. Every emitter must
-    therefore declare `suppresses` only for a code it directly explains,
-    never for one that suppresses something else in turn.
-    """
-    suppressed = {code for d in diagnostics for code in d.suppresses}
-    return [d for d in diagnostics if d.code not in suppressed]
-
-
-def count_by_severity(diagnostics: list[Diagnostic]) -> dict[str, int]:
-    return {
-        "errors": sum(1 for d in diagnostics if d.severity == SEVERITY_ERROR),
-        "warnings": sum(1 for d in diagnostics if d.severity == SEVERITY_WARNING),
-    }
 
 
 # --- spec -------------------------------------------------------------------
-
-
-@dataclass
-class Node:
-    id: str
-    label: str
-    service: str | None = None
-    sublabel: str = ""
-    zone: str | None = None
-    color: str = "#3A3A3A"
-    provider: str = "generic"
-    owner: str | None = None
-    external: bool = False
-    storage: bool = False
-
-
-@dataclass
-class Zone:
-    id: str
-    label: str
-    parent: str | None = None
-    kind: str = "generic"
-
-
-@dataclass
-class Edge:
-    src: str
-    dst: str
-    label: str = ""
-    type: str = "default"
-    id: str | None = None
-
-
-@dataclass
-class Spec:
-    title: str
-    direction: str
-    provider: str
-    nodes: list[Node]
-    zones: list[Zone]
-    edges: list[Edge]
-    profile: str | None = None
-
-
-class SpecError(ValueError):
-    """A spec the renderer cannot answer.
-
-    Carries the `Diagnostic` so callers get the same coded record they get
-    for every other finding; the exception message stays human-readable for
-    tracebacks.
-    """
-
-    def __init__(self, diagnostic: Diagnostic) -> None:
-        super().__init__(diagnostic.message)
-        self.diagnostic = diagnostic
-
-
-def _spec_error(
-    code: str,
-    message: str,
-    subject: dict[str, object] | None = None,
-    evidence: dict[str, object] | None = None,
-    supported_fixes: tuple[str, ...] = (),
-) -> SpecError:
-    return SpecError(
-        Diagnostic(
-            code=code,
-            severity=SEVERITY_ERROR,
-            message=message,
-            subject=subject or {},
-            evidence=evidence or {},
-            supported_fixes=supported_fixes,
-        )
-    )
-
-
-def load_spec(text: str) -> Spec:
-    data = yaml.safe_load(text) or {}
-    if "nodes" not in data or not data["nodes"]:
-        raise _spec_error(
-            "spec/no-nodes",
-            "spec has no nodes",
-            supported_fixes=("add at least one entry under `nodes`",),
-        )
-
-    provider = data.get("provider", "generic")
-    profile = data.get("profile")
-    if profile not in (None, "deployment-ownership"):
-        raise _spec_error(
-            "spec/unknown-profile",
-            f"unknown profile: {profile!r}",
-            evidence={"profile": profile},
-            supported_fixes=(
-                "omit `profile` for the default visual-only validation",
-                "set `profile: deployment-ownership`",
-            ),
-        )
-
-    node_ids = set()
-    nodes = []
-    for raw in data["nodes"]:
-        if "id" not in raw:
-            raise _spec_error(
-                "spec/node-missing-id",
-                f"node missing id: {raw}",
-                evidence={"node": raw},
-                supported_fixes=("give the node a unique `id`",),
-            )
-        for field_name in ("external", "storage"):
-            value = raw.get(field_name, False)
-            if not isinstance(value, bool):
-                raise _spec_error(
-                    "spec/invalid-node-boolean",
-                    f"node {raw['id']!r} field {field_name!r} must be boolean",
-                    subject={"node": raw["id"]},
-                    evidence={"field": field_name, "value": value},
-                    supported_fixes=(
-                        f"set `{field_name}` to true or false without quotes",
-                    ),
-                )
-
-        if raw["id"] in node_ids:
-            raise _spec_error(
-                "spec/duplicate-node-id",
-                f"duplicate node id: {raw['id']}",
-                subject={"node": raw["id"]},
-                supported_fixes=(
-                    "rename one of the nodes so every `id` is unique",
-                    "delete the duplicate node entry",
-                ),
-            )
-        node_ids.add(raw["id"])
-        nodes.append(
-            Node(
-                id=raw["id"],
-                label=raw.get("label", raw["id"]),
-                service=raw.get("service"),
-                sublabel=raw.get("sublabel", ""),
-                zone=raw.get("zone"),
-                color=raw.get("color", "#3A3A3A"),
-                provider=raw.get("provider", provider),
-                owner=raw.get("owner"),
-                external=raw.get("external", False),
-                storage=raw.get("storage", False),
-            )
-        )
-
-    zone_ids = set()
-    zones = []
-    for raw in data.get("zones", []) or []:
-        if "id" not in raw:
-            raise _spec_error(
-                "spec/zone-missing-id",
-                f"zone missing id: {raw}",
-                evidence={"zone": raw},
-                supported_fixes=("give the zone a unique `id`",),
-            )
-        zone_ids.add(raw["id"])
-        kind = raw.get("kind", "generic")
-        if kind not in ("generic", "region", "security"):
-            raise _spec_error(
-                "spec/invalid-zone-kind",
-                f"zone {raw['id']!r} has invalid kind {kind!r}",
-                subject={"zone": raw["id"]},
-                evidence={
-                    "kind": kind,
-                    "allowed_kinds": ["generic", "region", "security"],
-                },
-                supported_fixes=(
-                    "omit `kind` for a visual-only generic zone",
-                    "set `kind` to `generic`, `region`, or `security`",
-                ),
-            )
-
-        zones.append(
-            Zone(
-                id=raw["id"],
-                label=raw.get("label", raw["id"]),
-                parent=raw.get("parent"),
-                kind=kind,
-            )
-        )
-    for n in nodes:
-        if n.zone is not None and n.zone not in zone_ids:
-            raise _spec_error(
-                "spec/unknown-zone",
-                f"node {n.id!r} references unknown zone {n.zone!r}",
-                subject={"node": n.id},
-                evidence={"zone": n.zone, "known_zones": sorted(zone_ids)},
-                supported_fixes=(
-                    "declare the zone under `zones`",
-                    "point the node's `zone` at an existing zone id",
-                    "drop the node's `zone` field",
-                ),
-            )
-    for z in zones:
-        if z.parent is not None and z.parent not in zone_ids:
-            raise _spec_error(
-                "spec/unknown-zone-parent",
-                f"zone {z.id!r} references unknown parent {z.parent!r}",
-                subject={"zone": z.id},
-                evidence={"parent": z.parent, "known_zones": sorted(zone_ids)},
-                supported_fixes=(
-                    "declare the parent zone under `zones`",
-                    "point `parent` at an existing zone id",
-                    "drop `parent` to make this a top-level zone",
-                ),
-            )
-        if z.parent == z.id:
-            raise _spec_error(
-                "spec/zone-self-parent",
-                f"zone {z.id!r} cannot be its own parent",
-                subject={"zone": z.id},
-                supported_fixes=(
-                    "drop `parent` to make this a top-level zone",
-                    "point `parent` at the enclosing zone",
-                ),
-            )
-
-    edges = []
-    for raw in data.get("edges", []) or []:
-        if raw.get("from") not in node_ids or raw.get("to") not in node_ids:
-            raise _spec_error(
-                "spec/unknown-edge-node",
-                f"edge references unknown node: {raw}",
-                evidence={"edge": raw, "known_nodes": sorted(node_ids)},
-                supported_fixes=(
-                    "point `from` and `to` at existing node ids",
-                    "declare the missing node under `nodes`",
-                ),
-            )
-        edges.append(
-            Edge(
-                src=raw["from"],
-                dst=raw["to"],
-                id=raw.get("id"),
-                label=raw.get("label", ""),
-                type=raw.get("type", "default"),
-            )
-        )
-
-    return Spec(
-        title=data.get("title", ""),
-        direction=data.get("direction", "LR").upper(),
-        provider=provider,
-        nodes=nodes,
-        zones=zones,
-        edges=edges,
-        profile=profile,
-    )
-
-
-_NODE_SEMANTIC_FIELDS = (
-    "label",
-    "service",
-    "sublabel",
-    "color",
-    "provider",
-    "owner",
-    "external",
-    "storage",
-)
-_NODE_SCOPE_FIELDS = ("zone",)
-_EDGE_SEMANTIC_FIELDS = ("label", "type")
-_EDGE_ROUTE_FIELDS = ("from", "to")
-
-
-def _comparison_error(
-    code: str,
-    message: str,
-    subject: dict[str, object],
-    evidence: dict[str, object],
-    supported_fixes: tuple[str, ...],
-) -> SpecError:
-    return _spec_error(code, message, subject, evidence, supported_fixes)
-
-
-def _canonical_nodes(spec: Spec) -> dict[str, dict[str, object]]:
-    nodes: dict[str, dict[str, object]] = {}
-    for node in spec.nodes:
-        if not isinstance(node.id, str) or not node.id:
-            raise _comparison_error(
-                "compare/invalid-node-id",
-                "comparison requires every node to have a non-blank string id",
-                {"node": node.id},
-                {"id": node.id},
-                ("set each node `id` to a non-blank string",),
-            )
-        nodes[node.id] = {
-            "label": node.label,
-            "service": node.service,
-            "sublabel": node.sublabel,
-            "color": node.color,
-            "provider": node.provider,
-            "owner": node.owner,
-            "external": node.external,
-            "storage": node.storage,
-            "zone": node.zone,
-        }
-    return nodes
-
-
-def _canonical_edges(spec: Spec) -> dict[str, dict[str, object]]:
-    edges: dict[str, dict[str, object]] = {}
-    for edge in spec.edges:
-        if not isinstance(edge.id, str) or not edge.id:
-            raise _comparison_error(
-                "compare/missing-edge-id",
-                "comparison requires every edge to have a non-blank authored id",
-                {"edge": {"from": edge.src, "to": edge.dst}},
-                {"id": edge.id},
-                ("set a unique non-blank `id` on every edge",),
-            )
-        if edge.id in edges:
-            raise _comparison_error(
-                "compare/duplicate-edge-id",
-                f"comparison found duplicate edge id {edge.id!r}",
-                {"edge": edge.id},
-                {"id": edge.id},
-                ("make every edge `id` unique within its specification",),
-            )
-        edges[edge.id] = {
-            "from": edge.src,
-            "to": edge.dst,
-            "label": edge.label,
-            "type": edge.type,
-        }
-    return edges
-
-
-def _field_changes(
-    base: dict[str, object], head: dict[str, object], fields: tuple[str, ...]
-) -> list[dict[str, object]]:
-    return [
-        {"path": f"/{field}", "base": base[field], "head": head[field]}
-        for field in fields
-        if base[field] != head[field]
-    ]
-
-
-def _compare_entities(
-    base: dict[str, dict[str, object]],
-    head: dict[str, dict[str, object]],
-    semantic_fields: tuple[str, ...],
-    relocation_fields: tuple[str, ...],
-    relocation_status: str,
-) -> list[dict[str, object]]:
-    changes: list[dict[str, object]] = []
-    for entity_id in sorted(base.keys() | head.keys()):
-        before = base.get(entity_id)
-        after = head.get(entity_id)
-        if before is None:
-            changes.append(
-                {
-                    "id": entity_id,
-                    "status": ["added"],
-                    "changed_fields": [],
-                    "before": None,
-                    "after": after,
-                }
-            )
-            continue
-        if after is None:
-            changes.append(
-                {
-                    "id": entity_id,
-                    "status": ["removed"],
-                    "changed_fields": [],
-                    "before": before,
-                    "after": None,
-                }
-            )
-            continue
-
-        semantic_changes = _field_changes(before, after, semantic_fields)
-        relocation_changes = _field_changes(before, after, relocation_fields)
-        status = []
-        if semantic_changes:
-            status.append("changed")
-        if relocation_changes:
-            status.append(relocation_status)
-        changes.append(
-            {
-                "id": entity_id,
-                "status": status or ["unchanged"],
-                "changed_fields": semantic_changes + relocation_changes,
-            }
-        )
-    return changes
-
-
-def compare_specs(base: Spec, head: Spec) -> dict[str, list[dict[str, object]]]:
-    """Compare canonical authored IR only; no generated geometry participates."""
-    return {
-        "nodes": _compare_entities(
-            _canonical_nodes(base),
-            _canonical_nodes(head),
-            _NODE_SEMANTIC_FIELDS,
-            _NODE_SCOPE_FIELDS,
-            "moved",
-        ),
-        "edges": _compare_entities(
-            _canonical_edges(base),
-            _canonical_edges(head),
-            _EDGE_SEMANTIC_FIELDS,
-            _EDGE_ROUTE_FIELDS,
-            "rerouted",
-        ),
-    }
 
 
 # --- ranking + ordering ------------------------------------------------------
